backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
python-multipart==0.0.9
mammoth==1.7.0
pdfplumber==0.11.0
elevenlabs>=1.50.0
boto3==1.34.51
python-dotenv==1.0.1
